+++ Source/GameEngine/Graphic/Scene/Element/Particle/ParticleMeshEmitter.cpp
@@ -76,13 +76,13 @@ int ParticleMeshEmitter::Emitt(unsigned int now, unsigned int timeSinceLastCall,
 						if( mMaxAngleDegrees )
 						{
 							Quaternion<float> tgt = Rotation<3, float>(
-								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 							particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(1)));
 							tgt = Rotation<3, float>(
-								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 							particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(0)));
 							tgt = Rotation<3, float>(
-								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+								AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 							particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(2)));
 						}
 
@@ -133,13 +133,13 @@ int ParticleMeshEmitter::Emitt(unsigned int now, unsigned int timeSinceLastCall,
 					particle.mVector = mDirection;
 
 					Quaternion<float> tgt = Rotation<3, float>(
-						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 					particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(1)));
 					tgt = Rotation<3, float>(
-						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 					particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(0)));
 					tgt = Rotation<3, float>(
-						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * GE_C_DEG_TO_RAD));
+						AxisAngle<3, float>(particle.mVector, Randomizer::FRand() * mMaxAngleDegrees * (float)GE_C_DEG_TO_RAD));
 					particle.mVector = HProject(Rotate(tgt, Vector4<float>::Unit(2)));
 				}
